refactor: 예약 취소를 canceler 컴포넌트가 아닌 service에서 하도록 변경 (피드백 반영)
